requirements.txt
discord
python-dotenv
tensorflow
numpy
scipy
opencv-python
pillow
matplotlib
h5py
keras
imageai